flaskapp/requirements.txt
docker==4.3.1
mysql-connector
docutils==0.15.2
Flask==1.1.2
Flask-RESTful==0.3.8
goodreads==0.3.2
Jinja2==2.11.2
jmespath==0.10.0
jsondiff==1.1.2
jsonpatch==1.26
jsonpickle==1.4.1
jsonpointer==2.0
jsonschema==3.2.0
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycparser==2.20
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
rauth==0.7.3
requests==2.22.0
requests-file==1.5.1
responses==0.12.0
SQLAlchemy==1.3.20
uWSGI==2.0.19.1
websocket-client==0.57.0
Werkzeug==1.0.1